Deprecate input validation
This commit removes array checking to accommodate for Scikit-learn's
greater flexibility in input arrays. It will be the user's
responsibility to pass viable input to the ensemble, and if not
the estimators will eventually raise an error.

This commit also remove the deprecated EnsembleTransformer.
