+++ mlens/ensemble/super_learner.py
@@ -113,19 +113,6 @@ class SuperLearner(BaseEnsemble):
         but estimation continues unless exception is fatal. Note that this
         can result in unexpected behavior unless the exception is anticipated.
 
-    array_check : int (default = 2)
-        level of strictness in checking input arrays.
-
-            - ``array_check = 0`` will not check ``X`` or ``y``
-
-            - ``array_check = 1`` will check ``X`` and ``y`` for
-              inconsistencies and warn when format looks suspicious,
-              but retain original format.
-
-            - ``array_check = 2`` will impose Scikit-learn array checks,
-              which converts ``X`` and ``y`` to numpy arrays and raises
-              an error if conversion fails.
-
     verbose : int or bool (default = False)
         level of verbosity.
 
@@ -213,7 +200,7 @@ class SuperLearner(BaseEnsemble):
 
     def __init__(
             self, folds=2, shuffle=False, random_state=None, scorer=None,
-            raise_on_exception=True, array_check=2, verbose=False, n_jobs=-1,
+            raise_on_exception=True, array_check=None, verbose=False, n_jobs=-1,
             backend='threading', model_selection=False, sample_size=20, layers=None):
         super(SuperLearner, self).__init__(
             shuffle=shuffle, random_state=random_state, scorer=scorer,
