+++ mlens/utils/__init__.py
@@ -11,27 +11,26 @@ from .utils import (
     kwarg_parser, clone_attribute)
 
 from .formatting import check_instances, format_name
-from .validation import check_inputs
 from .checks import (
     check_ensemble_build, assert_valid_estimator, assert_valid_pipeline,
     assert_correct_format, check_initialized)
 
-__all__ = ['IdTrain',
-           'check_inputs',
-           'check_instances',
-           'check_ensemble_build',
-           'assert_correct_format',
-           'assert_valid_estimator',
-           'assert_valid_pipeline',
-           'check_initialized',
-           'pickle_save',
-           'pickle_load',
-           'load',
-           'time',
-           'print_time',
-           'safe_print',
-           'CMLog',
-           'kwarg_parser',
-           'clone_attribute',
-           'format_name'
-           ]
+__all__ = [
+    'IdTrain',
+    'check_instances',
+    'check_ensemble_build',
+    'assert_correct_format',
+    'assert_valid_estimator',
+    'assert_valid_pipeline',
+    'check_initialized',
+    'pickle_save',
+    'pickle_load',
+    'load',
+    'time',
+    'print_time',
+    'safe_print',
+    'CMLog',
+    'kwarg_parser',
+    'clone_attribute',
+    'format_name'
+]
